+++ docs/release-notes.md
@@ -4,6 +4,8 @@
 ## Upcoming released
 * For modders:
   * Added `Multiplayer.PeerConnected` event.
+  * Migrated to Harmony 2.0 (see [_migrate to Harmony 2.0_](https://stardewvalleywiki.com/Modding:Migrate_to_Harmony_2.0) for more info).
+  * Harmony mods which use the `[HarmonyPatch(type)]` attribute now work crossplatform. Previously SMAPI couldn't rewrite types in custom attributes for compatibility.
   * Fixed asset propagation for Gil's portraits.
 
 ## 3.5